Add support for new primitive to handle the late-known sync/single fields
Current implementation copies a lot from InitNormalize's insertion.  It would
be good to find a way to unify this implementation to avoid code drift.

This solution was able to also handle the earlier fix for ordinary sync fields
without a type, so remove the semi-mystical implementation of that and allow it
to follow the same pathway as when the type is not yet determined.