Make typeForTypeSpecifier() able to squash errors
Added a new 'fatalOK' argument to typeForTypeSpecifier() that
indicates whether it is reasonable for the routine to issue a fatal
error when the type expression is found to be illegal (e.g.,
'int(3)').  Calls early in the compiler working on code that may later
be folded out or be found to be unreachable should set this to 'false'
to avoid issuing errors prematurely.