Add a bash autocompletion script for chpl options
Sourcing this script enables tab-autocompletion for Chapel compiler options in
bash. It contains a static list of compiler options. The commands used to
generate the lists are included as comments. It should be possible to generate
these lists on the fly, but I didn't want to require a working chpl to be in
the path.