Convert locale models over to use init() rather than constructors
This change converts locale models over to use init() rather than
constructors.  For the most part this was straightforward, though for
the current compiler the 'parent' field seems to only be able to be
initialized in phase 2 (in spite of seemingly being 'const'?) and this
required using the `_parent` formal argument in a few cases that
wanted to use the field in phase 1 prior to its being initialized.