Adjust test to work with locale model init()s
The introduction of init() routines in locale models causes the output
of this test to change.  Here, I'm taking Lydia's suggestion to update
the .good file to match the expected output of the 'flat' locale model
(our common case) and to suppress the output for other locale models.
When we do a better job of filtering out method calls whose base type
doesn't match that of the object on which the method was dispatched,
this suppression should be able to be removed (as well as the new
output in the .good file).