Add a script to generate the chpl-completion.bash script from a template
Add a script that uses `chpl --devel --help` to generate the bash
autocompletion script from a template.  Add a test that uses the new script
to generate the autocomplete script and diff it against the committed
version.  If the test finds any differences it probably means a compiler
option was added or removed, so the autocompletion script should be
regenerated.