Add two tests of this.init() calls for generic records
These tests exercise the bug revealed by #7960, where a this.init() call in a
generic record causes an internal error in initializerResolution.cpp, due to
failing to resolve the call.  I'm uncertain of the level of effort to fix this
bug, but I strongly suspect it will not be trivial.

One is a copy of the original test, the other is a copy of a simpler test that
I wrote when resolving this issue for classes.  Both tests pass when a class is
defined instead of a record, with no further modifications.