Replace 'proc these' in ChapelArray with forwarding iters
The 'proc these' that returns an iterator record causes problems
for iterators that include default arguments, because these default
arguments will be (after 7858) allocated at the call site (before
that PR it is worse). The call site is 'proc these', which will
return before the iterator runs. That's a problem if the iterator
captures these default arguments by reference.