Adjust inlineFunctions to inline QUAL_REF args as QUAL_REF tmps
Before this change, it was possible for the inliner to create AST that might
have new meaning. Note first that PRIM_GET_MEMBER and PRIM_SET_MEMBER might not
be valid depending on the ref-ness of the field and value.

In particular, the inliner might have seen this pattern:

  record R { var x:int; }

  inline proc f(arg:R) {
    ... PRIM_SET_MEMBER someRefField, arg ...
  }

  proc test() {
    var loc:R;
    f(loc);
  }

The AST for this pattern before inlining involves a QUAL_CONST_REF argument
to f but just the value 'loc' passed to it directly, without any indication
that an addr-of operation occured at the call site. (This is by design for
QualifiedTypes and PRIM_ADDR_OF).

For example, the AST dumped just before inlining might be:

  function f(const ref arg arg:R) : void
  {
    PRIM_SET_MEMBER someRefField, arg
    return _void
  }

  function test() : void
  {
    val loc:R
    ...
    (call f loc)
    return _void
  }

Now, supposing that 'f' used PRIM_GET_MEMBER or PRIM_SET_MEMBER, the
ref-ness of 'arg' in the inlined code might change the meaning. Therefore,
this commit changes the inliner to create a reference temporary in the
event that the formal argument is a reference but the actual argument is not.

I.e. before this commit, the inliner would produce something like this for the
above example:

function test()
{
  val loc:R;
  ...
  PRIM_SET_MEMBER someRefField, loc
  return _void
}

and in particular, this is invalid AST according to the current rules for how
PRIM_SET_MEMBER can interact with reference fields.

But after this commit, it will produce:

function test()
{
  val loc:R;
  ...
  const ref tmp:R = PRIM_SET_REFERENCE(loc)
  PRIM_SET_MEMBER someRefField, tmp
}

which is valid AST and indicates the same behavior as the non-inlined version.

This commit preserves the ref-ness of the replaced formal arguments in the
copied function body so that the AST does not change meaning in the event that
operations have different meaning depending on ref-ness (as is the case with
PRIM_SET_MEMBER and PRIM_GET_MEMBER).