Adjust addVarsToActuals to handle Qualified-Ref fields
In flattenFunctions, addVarsToActuals updates call sites with actual arguments
that refer to outer variables. In the event that the argument should be "passed
by reference", it previously would always add a PRIM_ADDR_OF call. However,
with Qualified Types, it can pass a local variable to a ref-intent actual
without creating temporary storing the result of PRIM_ADDR_OF.  That is
important in the event that the formal argument is passed by ref but does not
use a ref type, as is currently the case with blank intent record arguments.