Introduce task-private variables (TPV)
This is the first step in implementing all we want/need
about task-private variable.

A task-private variable is similar to a forall intent:

    forall idx in myIter() with (var myTPV = 5) ...

Currently only "var VARNAME = INIT EXPR" form is supported.

FUTURE WORK

* other qualifiers: const, ref, const ref

* allow declaring the type of the TPV, with or without
  the initialization expression

* support for the case where the parallel iterator being invoked
  has a yield in a forall of parallel-for loop;
  ex. for now can't use TPVs in a forall over an array

* TPV init expression to reference a reduce-intent shadow variable
  (support for this is farther out in the future; it may just happen
   if we make a certain change to the representation)

* task startup and tear-down blocks (need syntax for these!)

* emit a useful error message upon a reference to a loop index variable
  within the task startup or tear-down block

* switch ShadowVarSymbol::isConstValWillNotChange() to using
  FLAG_REF_TO_IMMUTABLE

COMPILER REPRESENTATION

* a ShadowVarSymbol
 - handled similarly to other forall intents

* a VarSymbol
 - is defined and initialized in the task init block
 - serves as the outerVarSym() for the ShadowVarSymbol

* taskInit block
 - for now, keeps only the defs and inits for the above VarSymbols
 - could also contain user-specified task startup code

* taskDeinit block
 - always empty for now
 - could contain user-specified task tear-down code

* within taskInit and taskDeinit lexical scopes:
 - a reference to a TPV refers to the above VarSymbol
 - a reference to a non-TPV outer variable refers
   to the corresponding ShadowVarSymbol
 - cannot reference a loop induction variable

* the loop body can access all variables declared in taskInit
  (currently, only TPVs), plus shadow variables and inductionvariables

* the lexical scope of taskDeinit is nested within taskInit

Conceptually, a ShadowVarSymbol has been strictly a creature
of the body of the forall loop. This change also allows it
to be used in taskInit/taskDeinit blocks. ShadowVarSymbols
for TPVs and non-TPVs are handled very similarly, see next.

Recall that a key part of the existing implementation of forall intents
is `extendLeaderNew()`. It extends the parallel iterator to propagate
outer variables through its task-parallel and data-parallel constructs
and yield them as part of a newly-created tuple. (We have been considering
getting away from this approach.)

This change keeps this implementation and extends it for TPVs.
Namely:

* Each task construct (lexically present in the parallel iterator)
is extended to execute a clone of taskInit at the beginning
and a clone of taskDeinit at the end.

* Recall that for a given forall intent the yielded tuple
contains the corresponding formal of the enclosing task construct.
(There is a bit more for a reduce intent.)

* For a TPV, there is no corresponding formal.
This is because the TPV starts its lifetime inside the task construct.
Instead, the role of the formal is played by the clone of
the corresponding VarSymbol that is a part of the clone of taskInit
that is executed at the beginning of the task construct (see above).

When the parallel iterator contains a forall loop:

* For each ShadowVarSymbol in the original forall loop,
a corresponding new ShadowVarSymbol is added to shadowVariables()
of the forall in the iterator.

* New clones of taskInit and taskDeinit of the original forall loop
are added to taskInit and taskDeinit blocks of the forall in the iterator.

(This handling of foralls in parallel iterators is to be implemented.)

WHILE THERE

Improved list_view() of ForallStmts.

Fixed ShadowVarSymbol::isConstValWillNotChange().

Minor refactoring. Also, moved bodies of two functions out
from QualifiedType class declaration.