Don't try to use "draining" nodes in paratest.chapcs
Draining nodes are counted as "allocated" so we're still counting them as an
"online" node, but we can't actually run anything on a draining node. Only the
currently running job can run, and then it will be marked as "drained".

To avoid this, just consider draining nodes as exclusive, and don't try to
reserve them.