Fix the insertion of omitted initialization for fields using cond/loop exprs
Made updateFieldsMember check if what it was provided was a call that turned
out to be a method.  If so, it alters the AST so that we can recognize it as
a method call later.  Then, if we recognize it is a method call and the method
is an inserted conditional or loop expression function, don't generate our
method warning.

The AST set up I used for this was a bit different from what normally is
present prior to normalize's applyGetterTransform().  It seemed silly to
insert a SymExpr to a VarSymbol with an Immediate storing the name that was
just going to get transformed back into an UnresolvedSymExpr when we already
had an UnresolvedSymExpr.  So I added a case to applyGetterTransform to handle
this new set up.