file.check throws; fix normalization; SystemError default args (#7908)
[reviewed by @mppf]

file.check() was halting when it should throw. It has been updated to throw a SystemError(EBADF), and file.check(out error) was added to help I/O functions transition.

file.check(out error) revealed errors with the lack of default initialization for extern types, which was worked around in the PR and filed under issue #7992.

Further testing revealed that file.check() was causing un-normalized code to be emitted by the errorHandling pass, leading to errors when programs were compiled with --no-local. This was addressed in the PR by modifying the output and use of castToError().

While there, it was found that subtypes of SystemError could not be built easily with just a string due to type coercions with syserr. To simplify this, a syserr cast was added to each initializer's default syserr argument.