Improve registration broadcast.
The main change here is in broadcasting new registrations.  When a new
entry extends the active range of the table of registered regions, we
used to broadcast the entire prefix of the table, from the count member
all the way through the new entry.  Now we broadcast just the entries
that are beyond what has been broadcast previously, and then we
broadcast the new count.  This reduces the amount of data sent for
memory region table updates but doesn't significantly improve overall
performance as far as I know, presumably because the amount of data
being transferred at current scales isn't enough to matter.

Also, when a registered memory region table entry is freed, we now zero
its GNI MDH along with its address and length.  This is really just to
make sure we don't end up broadcasting expired MDHs when we reuse an
entry after it's freed.  We hadn't seen any failures that seemed to be
caused by this, but doing it has essentially no cost in this context and
seems prudent.

Finally, here we also change a check in fork_call_common() that tries to
save some overhead by pre-duplicating large executeOn arg bundles into
registered memory if they aren't there already.  (Because if they aren't
then when the target node creates its buffer and tries to GET them, the
GET will end up having to executeOn back to the initiating node and PUT
them from there.  So if pre-duplicating puts the copy in registered
memory, we save an executeOn.)  The old logic just checked whether we
had a registered heap.  With dynamic registration that no longer covers
all cases, so the new check is just whether we have hugepages.  Really
what we want to know is: if we do a chpl_mem_allocMany(), will we get
registered memory?  We can't ask that directly, but checking for
hugepages is a pretty good proxy.

In terms of lines changed, the biggest items are really while-I-was-here
ones: the vector broadcast code for memory region table updates is now
in a little function called from regMemPostAlloc() and regMemFree(), and
I made improvements to much of the commentary having to do with region
table updates.