Permit copy initializers to take the 'other' argument by ref
The BigInt type had a copy initializer, yet it was not being
recognized as such because it took its 'other' argument in by 'const
ref' intent.  This change permits such intents to also be matched
when looking for the copy initializer.

Without this, the following three tests failed when run with a version
of the modules that converted constructors to initializers:

* library/standard/GMP/ferguson/gmp_random.chpl
* library/standard/GMP/ldelaney/gmp_Bigint_record_constructors.chpl
* library/standard/GMP/ldelaney/gmp_random_2.chpl