Make the the `tzinfo` time zone field in DateTime.time use Shared
Make the `tzinfo` field in DateTime.time Shared so that it will be
automatically deleted when there are no more references to it.  Updated
functions/methods that have a `TZInfo` argument to use `Shared(TZInfo)`
instead.  Added calls to `borrow()` as needed, most commonly when comparing
it with `nil`.

Updated DateTime tests to use `Shared(TZInfo)` instead of `TZInfo` directly.
One pain point is that real uses of time zones use a subclass of `TZInfo`
and need to cast it to the `TZInfo` type before making it shared.

  var sharedTZ = new Shared(new DerivedTZ(...): TZInfo);