Add an llvm --no-local performance configuration
--no-local prk-stencil performs much worse with the llvm backend, so I want to
see if there's other regressions as well.

This will run on the same machine as regular --no-local perf testing, which
required moving clang to a different machine, so drop clang to 1 trial (it's
not a configuration we care deeply about) and only do 3 trials for --no-local
perf configs since they're a little slower.