Rely on externally-set CHPL_AUXIO* Make variables to provide lustre api.
This change simply reverts PR #3550. PR #3550 was not entirely successful
because it only fixes the "missing file lustre/lustreapi.h" problem for
Cray-PrgEnv compilers- it does not help the clang-included compiler, which
does not interact with the Cray PrgEnv at all.

As we have recently decided to use the CHPL_AUXIO* Make variables to
provide lustre api paths and flags for clang-included, we may as well let
the the Cray-PrgEnv compilers get the same info the same way.