Follow type aliases to work around resolution order issues
The removal of c_string_copy (#8014) resulted in a failure for the
following test:

extern/jjueckstock/basic.chpl

That PR adjusted the _defaultOf for c_string to return a cast from
'c_nil' to 'c_string'. This requires the resolution of a '_cast'
function in CPtr.chpl with a where-clause of:

```chpl
(t:c_intptr || t:c_uintptr || t:int(64) || t:uint(64)) &&
(x.type:c_void_ptr || x.type:c_ptr)
```

The compiler hadn't yet resolved the 'c_intptr' type, and so the
resolution of the where-clause failed. This commit updates postFold's
handling of PRIM_IS_SUBTYPE to follow type aliases, which correctly
finds the type of c_intptr and allows the where-clause to resolve.