Compiler changes to fix internal failure with autoCopies when returning record
Prior to this change, returning an argument whose type was a record that defined
an initializer would result in an internal compiler error about autoCopy,
regardless of whether a copy initializer was defined on the type or not.  We
wanted the compiler to generate an error message when a copy initializer was
not defined, and for compilation to succeed when one was present.

This change follows option 2 in issue #8064 (allow generic, single-argument
initializers to be recognized as copy initializers, but generate an error
message if that fails to resolve correctly, indicating that the function is
being treated as a copy initializer when perhaps it shouldn't be).

Though this fixes the afore-mentioned failures, it causes an array test case to
regress, as it has not defined a copy initializer explicitly.  Our next change
will be to generate copy initializers by default when one is not provided.