Add tests of returning an iterator expression
Now that returning an iterator turns it into an array first,
the RandomNumber.chpl test would pass. But it's still possible
to get the old behavior and the old error message with a pragma.
So I replaced RandomNumber.chpl with two tests that check for
these errors - one that uses the pragma.