Improve memory semantics in DynamicIters
* Pass `remain` and `r` ranges into `coforall` by `const in`
instead of by `ref`.

As DavidK says, this will avoid potential unnecessary cache line
sharing between cores. Also, passing `remain` by `ref` is illegal
because `remain` is a `const`. Even though the compiler did not complain.

Note that in the guided iterator we must pass `remain` by `ref`
because it is passed to and modified within `adaptSplit()`.

* Switch `moreWork` and `undone` to atomics.

`moreWork` can be written to concurrently, i.e. without synchronization.
Both `moreWork` and `undone` can be read from concurrently with
being written to.

As DavidK says, while this may be safe without atomics on x86,
it may not be the case for other architectures.

A change in `adaptSplit()` is made to allow the formal to be
either an atomic or not.

* Removed a couple of explicit task intents in `with` clauses
because the implicit intents do the right thing already.