(Partial?) fixes for generic record copy initializers
Allow us to generate a copy initializer for generic records, mark all generated
copy initializers as FLAG_DEFAULT_COPY_INIT and use that flag in
computeSubstitutions so that the compiler doesn't confuse them with the default
initializer.  Extend the error message in resolveInitField so that it also
informs the user if the error occurred because we were trying to resolve
something as a copy initializer.

There may be more errors to fix, but this allows us to update the failing
generic record test reuse-init-new-instantiation-var so that it succeeds again.