Stop using point-of-instantiation in isMoreVisible
Function disambiguation rules use a function, isMoreVisible, to tell when one
function is more visible than another. This function has historically treated
the point of instantiation of a generic as the "parent block" for that
function. However that leads to confusing program behavior in some cases,
as descibed in CHIP 20.