Memory registration tuning improvements.
Reduce the amount of work done in memory registration single-threaded
regions.  A new atomic variable counts the number of free entries in the
local memory region table.  In the registered memory allocation function
this allows us both to check whether or not there is a free table entry
to hold the region and to do the hugepage allocation before acquiring
the single-threading mutex.  In the post-allocation function the memory
registration call is now done before acquiring the mutex, and in the
free function the deregistration call is now done before acquiring it.
I also switched from a forward to a reverse search for the table entry
in the registered memory past-alloc (registration) and free functions,
under the supposition that regions being registered and freed are more
likely to be dynamic ones that were recently acquired, rather than the
static ones that occupy the leading positions in the table.

Add short-circuit code to the memory region lookup functions, which
first check to see if the search address is in the same region returned
last time on this same thread.  This reduces the number of comparisons
done for lookups by about 20% for ISx, and the amount of time spent
doing lookups by about 10%.  (But the effect on overall performance is
less than this because such lookups are not a large proportion of the
execution time.)

There are a number of internal performance statistics changes here which
support the other work.  I added counters for memory region update
broadcasts and the table entry comparisons in memory region searches.  I
also added the ability to compute elapsed times, and then sums of the
time spent in memory region lookups and single-threaded in memory region
table updates.

While I was here, I removed a sanity check regarding the order of the
members in the memory region table descriptor struct.  This is no longer
needed because we don't transfer the count and table entries in the same
transaction any more.  I missed removing this in the last commit, which
changed how these were transferred.

Finally, I also added a missing right-paren that resulted in a syntax
error only when debug was enabled.