Make the condFieldType test a future again
Something about the default generated copy initializer is running into problems
with types that are determined via conditionals.  I'm not sure why this is, but
I suspect it is a thinko, albeit one that is more involved to track down.
Disable the test again for now, and come back to it later