forward RandomStream.iterate() leader to PCGRandomPrivate leader
Our standard Random module already forwarded NPBRandomStream.iterate()
to NPBRandomPrivate leader. This change adds that for PCG.

While there, reordered the 'tag' formal to follow our convention
that it is the last formal in leader iterators.

Apparently the only test we had that relied on this forwaring for NPB
was test1D2D. Here I extended it to test forwarding for PCG as well.
For that, I created a separate trivial wrapper. This is because
the outputs for comm=none and multilocale are different and
I needed both outputs for each rng algorithm.
I accepted, without checking, the current output of the PCG test
(for both single- and multi-locale runs) to be the correct output,
as we test PCG correctness elsewhere.

Trivial, not reviewed.
Thanks to @mppf for suggesting this change.
@mppf also suggests that we can do a better job of testing
Random's parallel iterators.