Minor refactorings/touch-ups
These are changes that should not change the compiler behavior in any way.
I made them "while there" - while working on the changes
in the upcoming commits.

* In CallExpr::verify, generalize the test sym == gMethodToken
to sym->type == dtMethodToken, then assert that indeed sym == gMethodToken.
This is to tighten our understanding of what to expect, currently.

* Rename pruneThisArg to pruneOuterVars. Because I am going to prune more.

* Refactor isOuterVar() and its use in findOuterVars(), for clarity.
Akin to what I did in their counterpart in implementForallIntents earlier.

* Replace collect_asts() with collectSymExprs() because we are only using
SymExprs. Get rid of the "is this a SymExpr" conditional.