DO exclude global const variables
As this test shows:

    optimizations/remoteValueForwarding/serialization/domains.chpl

esp:

    const globalDom = {1..10, 1..10};

    proc main() {
      coforall loc in Locales do on loc {
	... globalDom ...
      }
    }

when we handle global consts with task intents, their accesses
within the task function no longer benefit from broadcasting,
causing an increase in communication. Therefore we exclude those.

Note that here are we are walking on thin ice because we do not
necessarily know the concrete intent and the type of the global.
See the comment in the code for more discussion.