Handle break/continue within forall loops
Previously, break and continue statements (without a specified loop label)
inside a forall loop resulted in vague compiler diagnostics.
With this change, a break statement results in a clear error
and a continue statement is properly implemented.

Implementation is straightforward,  following the treatment of
break/continue for LoopStmt. One thing is of note:

I tried first to extend the existing LoopStmt::findEnclosingLoop() functions
to account for enclosing ForallStmt loops as well. That did not go well, however,
as it was unclear how to update most of the existing callers of those functions.
So instead I added findEnclosingLoopOrForall when the caller wants to know both.
I am leaving it as future work to figure out what the existing findEnclosingLoop
functions should do when the nearest loop is a forall.

While there, allowed multiple "bad break/continue/label" errors in one
compiler execution and show the label name in those error messages.