Add initialization of parSafe
It turns out that the error I was getting in my previous attempts to
set parSafe were due to initializing it in the wrong order because
I was following the initializer argument order rather than the field
order (and because the error message wasn't particularly clear -- see
test/classes/initializers/phase1/unorderedFields-preferred.future).
Here, I fixed that by re-arranging the field order to match the
initializer order because that seemed the least confusing / most
consistent.