Unify AggregateType::getFieldPosition behavior with the getField method
getFieldPosition returns the field of an AggregateType using 0-based indexing.
However, the getField method relies on 1-based indexing.  We don't use the
former at all in the compiler, but I wanted to use it for my initializers
implementation and it would be useful to have those methods use the same
indexing, even though the functionality of

getField(getFieldPosition("foo", false)

is already handled by

getField("foo", false)