Convert LinearAlgebraJama package module to initializers
The old constructor-based scheme used to initialize fields
in fairly arbitrary order, so the main effort here was
reordering things and teasing phase 1 away from phase 2.

Most of the constructors assigned m and n (the matrix sizes)
first, so I moved those fields higher up in the object
structure rather than trying to write workarounds to
initialize them later (this also matched a few cases that
did declare them first).  As a result of this reordering,
the default class print output changed, requiring an update
to the .good file, but this still felt better than trying to
delay initialization of m and n.