Avoid yielding during dynamic memory allocation/registration under ugni
This fixes the oversubscribed array allocation test, and will be important for
some upcoming fully dynamic registration work. We could have fixed the
oversubscribed array allocation test by using a chpl_sync_var as a lock or by
using a try-lock pthread_mutex, but this is the "better" long term solution.
For full dynamic memory registration we need to ensure that an allocation
finishes before a task yields because some allocations in qthreads (and the
chunk hooks in jemalloc) can't handle concurrent/interrupted allocations on the
same pthread.

Note that this hurts ISx performance. On 16 node this takes us from ~8s to ~11s