Add 6 tests to track the bug Russel encountered earlier with bigint
These cover modifications that do and do not reveal the bug found.

reductionOverCreatedArray-zero.chpl is a simplified version of the bug, showing
that the problem occurs when the index scope of the array created is 1..0

reductionOverCreatedArray-arrayTemp.chpl is the simpliest modification to allow
the program to work correctly.  It saves the created array into a temporary
before using it in the reduction.

reductionOverCreatedArray-generalType.chpl is my attempt to mimic bigint so that
I could get a more general case.  It does not encounter the bug, indicating that
one of the additional qualities of bigint is triggering this issue.

reductionOverCreatedArray-reductionTemp.chpl,
reductionOverCreatedArray-reductionTemp2.chpl, and
reductionOverCreatedArray-reductionTemp2a.chpl are variations involving saving
the result of the reduction operation with the array creation prior to returning
it.  reductionOverCreatedArray-reductionTemp2 demonstrates that in developer
mode, the bug is avoided for a single statement, while
reductionOverCreatedArray-reductionTemp2a shows that --no-devel will still
demonstrate the bug.  reductionOverCreatedArray-reductionTemp shows that even
though developer mode helps the single call case, it does not help the case
where it is called from a loop.

Save these cases while I continue to investigate what is causing the issue.