Output time information to stdout not stderr
Otherwise we see sporadic failures with
  studies/shootout/fasta/kbrady/fasta-printf
in a configuration with a launcher.

Adding a sys.stdout.flush() beforehand didn't solve the problem either.
I suspect that there's a launcher or other multi-node software in the way
of the flush strategy working.

I think it's reasonable to always output the time on stdout because if a
perf test is outputting anything to stderr it's probably failing.