Fix elegant CoMD and file future against blocking issue
With DateTime converted to initializers, CoMD started failing in a
surprising way due to its local-scope 'use' of DateTime which
prevented the copy initializer from being found.  With a quick
effort, I was unable to boil this down to a simple case for
Michael and Lydia to look at, so with Ben's permission, I instead
took the approach of moving the 'use' out to file scope which
permits it to work and filed a new future against the failure mode.
I'll open an issue against it as well.