Add a compilerError when trying to use the GMP module with CHPL_GMP=none
Currently, if you try to use the GMP module and CHPL_GMP=none, you'll get
non-obvious errors from the backend compiler about implicit declarations of GMP
types and GMP functions. This is especially confusing with speculative builds
of GMP, where the build failure is hidden and CHPL_GMP is silently set to none.
As an improvement, give an explicit compilerError that GMP support is not
available if CHPL_GMP=none.