exclude throwing from defer, deinit() (#8161)
[reviewed by @mppf]

defer blocks and deinit functions are similar constructs, since both are set to run at the end of some context. On this basis, @mppf and I wanted them to have a similar policy when it comes to throwing errors. After some discussion, we determined that allowing errors to be thrown in these contexts would lead to confusing control flow and that they should be banned from doing so until a strong motivating case and a coherent strategy are found.

For defer it becomes confusing when a series of defer blocks are being executed, and then one throws -- it's unclear if the other defer blocks will be executed. For deinit, a throw midway through could prevent the deinit from finishing altogether, adding complications for users. In both cases, the context in which the error should be handled is not clear, given that both are usually called implicitly.