Future-ize this test
This test redefines _defaultOf for string. The current PR causes it to fail,
but given my understanding of current language design issues around
point-of-instantiation rules, strings created in generic functions implementing
writeln (say) will resolve to the _defaultOf here, which isn't reasonable.
Based on that argument and discussion with @lydia-duncan, future-izing it for
the time being.