Split error with iteratorWithOn.chpl into future
This PR causes a problem with inferring a serial iterator's yield
type in the event the only yields are within 'on' statements.
This commit adjusts the relevant test to try it both with and
without a specified return type, and the unspecified return
type version is a future for now.