Add alternate resolution path for calling initializers w/ generic instantiation
Prior to this change, we would fail to resolve initializer calls when the type
provided was a generic instantiation instead of the generic base.  This was
because we were recognizing it as non generic, but for generic classes we only
create the _new function after we have resolved an initializer for that type.

This change adds a new path to resolveNew, which distinguishes non-generic types
from generic instantiations.  It then attempts to resolve the initializer as
though the fully generic type was provided, then compares the determined
instantiation to the one that was initially provided.  If they don't match, then
the best initializer match for the call could not generate the instantiation
desired, so the call was incorrect.  Note that this strategy does not handle the
case where a field could be coerced into the proper type.

This change resolves the future
test/classes/initializers/generics/typeFunctions/fnReturnsConcrete and allows
fnReturnsConcrete2 to pass.  It still needs testing over the slew of
initializers tests.