Fix default arguments for init
After PR #8141, we were seeing failures in getInitialization
(which, as called in parallel.cpp, looks for the initialization point for global const variables). The AST looked something like this:

 DefExpr unixEpoch
 DefExpr def_tmp
 init_default_tzinfo( &unixEpoch, /*ret*/ &def_tmp)
 init( &unixEpoch, &def_tmp )

(Note that PR #7858 rewrote default argument handling to use
 functions returning the default value - instead of a wrapper
 function calling the entire original function).

In other words, the function producing the default value for
an argument to an init function was consuming the 'this' argument.
But that doesn't make sense because 'this' won't be initialized yet
(it's initialized in the initializer). Therefore, I extended logic
in buildDefaultedActualFn and createDefaultedActual that avoids
passing 'this' to the default-argument function to apply to init
functions in addition to constructors.