Update the checkCopyrights script
This change updates the script used to check source files for
Cray copyright notices (including the current year).

The list of files in previously-scanned toplevel subdirs is
adjusted to reflect the kinds of "source" files they now contain,
or seem most likely to be added from now on.

Recently-added top-level subdir "tools" will now be scanned- before
this change, "tools" was not scanned at all.

The regexp used to check for valid copyright year is slightly modified.
Now, the LAST 4-digit year in the "Copyright ... Cray Inc" line must
be the current year. For ex, "2018-2004" would no longer be accepted.